requirements.txt
Flask==2.1.0
python-pptx==0.6.22
python-dotenv==0.19.2
requests==2.26.0
openai
elevenlabs
